spiffs - Working

+++ main/include/app_cycle.hpp
@@ -2,10 +2,10 @@
 #pragma once
 #include "trajectory_buffer.hpp"
 #include "ml_inference.hpp"
-#include "../spi_interface.h"
+#include "spi_interface.h"
 
 
 namespace AppCycle {
     void setup();
-    void run(uint16_t cycles);
+    void run(uint32_t cycles);
 }
